+++ src/main/java/com/example/searchengine/indexing/IndexingService.java
@@ -5,12 +5,12 @@ import com.example.searchengine.config.Site;
 public interface IndexingService {
 
      void init();
-     boolean canStartIndexing(Integer id);
-     String getSiteUrlForId(Integer id);
+     boolean canStartIndexing(long id);
+     String getSiteUrlForId(long id);
      boolean isIndexingInProgress();
      void setIndexingStopped();
      Site generateNewSite(String url);
-     String getStatus(Integer id);
+     String getStatus(long id);
      void logInfo(String message, Object... args);
 
 }
